# frontend/openapi.json
{
  "components": {
    "schemas": {
...
      "DemoLogin": {
        "description": "デモでログインする人。名前だけで identity を作る。",
        "properties": {
          "name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "表示名。空なら既定の名前を使う",
            "title": "Name"
          }
        },
        "title": "DemoLogin",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "title": "未来のPRネタを作り出すAI",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/calendar/demo-login": {
      "post": {
        "description": "Google を通さず名前だけでログインする。初めての名前なら予定を積んで返す。\n\nArgs:\n    body: 表示名。空なら既定の名前を使う。\n    request: user-agent と接続元をセッションに残すために使う。\n    db: データベースセッション。\n\nReturns:\n    204。セッション Cookie を張る。\n\nRaises:\n    HTTPException: APP_ENV が demo でないとき。",
        "operationId": "demo_login_api_calendar_demo_login_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DemoLogin"
              }
            }
          },
          "required": true
        },
        "responses": {
          "204": {
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Demo Login",
        "tags": [
          "calendar"
        ]
      }
    },
...
    "/api/calendar/oauth/callback": {
      "get": {
        "description": "認可コードを交換し, ユーザーを作ってログインさせ, カレンダーのトークンを保存する。\n\nArgs:\n    request: state の Cookie とクライアント情報を読むために使う。\n    code: Google が返した認可コード。\n    state: Google が返した突き合わせ用の値。\n    db: データベースセッション。\n\nReturns:\n    フロントへのリダイレクト。セッション Cookie を載せる。\n\nRaises:\n    HTTPException: state が一致しないとき 400。交換や ID token の検証に失敗したとき 502。",
        "operationId": "oauth_callback_api_calendar_oauth_callback_get",
        "parameters": [
          {
            "in": "query",
            "name": "code",
            "required": true,
            "schema": {
              "title": "Code",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "state",
            "required": true,
            "schema": {
              "title": "State",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Oauth Callback",
        "tags": [
          "calendar"
        ]
      }
    },
...
  }
}